register3 v1

+++ src/components/reinitialiser.vue
@@ -3,49 +3,53 @@
     <div :style="{'background-image': 'url(' + require('../assets/ingredient_fond.jpg') + ')'} " class="background_register1" >
     <div class="container">
         <div class="head">
-        <a href="/myaccueil"><i class="fas fa-arrow-circle-left"></i></a>
-        <h1>Nouveau <br>mot de passe</h1>
+        <a href="/"><i class="fas fa-arrow-circle-left"></i></a>
         </div>
+        <div class="section_centrale">
+        <h1>Nouveau <br>mot de passe</h1>
+        
         <div class="input_login">
 
-            <i class="fas fa-lock user"> <span></span>
-            <input type="password" name="password" id="password_register" placeholder="Password..." v-model="password_login" >
-            </i>
+            
+            <input type="password" name="password" id="password_register" placeholder="Password..." v-model="password_update" >
+           
 
-            <i class="fas fa-lock user"> <span></span>
-            <input type="password" name="confirm" id="confirm_register" placeholder="Confirmation password" v-model="confirm_register" >
-            </i>
+            
+            <input type="password" name="confirm" id="confirm_register" placeholder="Confirmation password" v-model="confirm_password_update" >
+            
 
             
         </div>
 
-       <button @click="connection" class="btn-valider"
+       <button @click="change" class="btn-valider"
             value="Valider">Valider</button>
+        </div>
     </div>
     </div>
   
 </template>
 <script>
 export default {
-    name: "login",
+    name: "reinitialiser",
     data() {
         return {
-            email_login: "",
-            password_login: ""
-            
+            password_update: "",
+            confirm_password: "",
+              
         }
 },
     components:{},
     methods:{
-        connection: function () {
-            this.axios.post("http://localhost:3000/user/login",{
-                email:this.email_login,
-                password:this.password_login
+        change: function () {
+            this.axios.post("http://localhost:3000/user/updatepassword",{
+                password:this.password_update,
+               confirm_password :this.confirm_password_update,
+                forget: this.$route.params.token
             })
             .then(res =>{
 
                 console.log(res);
-                alert("ok");
+                alert("Mot de passe changé");
                 if(res.data.token){
                     localStorage.setItem("token",res.data.token)
                     this.$router.push({name: 'register2'})
@@ -78,27 +82,33 @@ export default {
 .container{
     background-color: #f1f1f1;
     width: 41%;
-    height: 62%;
+   height: 27%;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     border-radius: 25px;
     position: relative;
+    box-shadow: 1px 1px 5px #585757;
 }
 
 .fa-arrow-circle-left::before{
-font-size: 45px;
-position: absolute;
-    left: 60px;
-    top: 58px;
+    font-size: 33px;
+    position: absolute;
+    left: 42px;
+    top: 37px;
     color: #a9a6a3;
 }
+.section_centrale{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 242px;
+}
+
 h1{
-    font-size: 60px;
-    text-decoration: underline #b0d18a;
+    font-size: 27px;
     text-align: center;
-    padding-bottom: 50px;
 }
 
 
@@ -109,48 +119,23 @@ h1{
     align-items: center;
     justify-content: center;
 }
-.input_login .user{
-    background-color: white;
-    width: 395px;
-    height: 45px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 1px solid #b0d18a;
-    border-radius: 5px;
-    margin: 20px;
-}
-.input_login .user input{
+input{
+    background-color: white !important;
     border: none;
-    font-size: 17px;
-    width: 243px;
+    font-size: 15px;
+    width: 260px;
     height: 33px;
-    margin-left: 8px;
-}
-
-.i{
-    margin: 10px;
-    height: 15px;
+    display: flex;
+    border-radius: 5px;
+    box-shadow: 1px 1px 5px #585757;
+    padding: 10px;
     
 }
-
 .input_login{
     display: flex;
     flex-direction: column;
     height: 34%;
     justify-content: space-between;
-}
-
-.fa-lock:before{
-font-size: 21px;
- color: #ff9d9e;
-}
-
-
-
-.fa-at:before{
-font-size: 21px;
- color: #ff9d9e
 }
 
 .mdp_forget{
@@ -160,7 +145,7 @@ font-size: 21px;
 }
 
 .btn-valider {
-    border: 0;
+   border: 0;
     height: 43px;
     width: 135px;
     font-size: 22px;
@@ -170,13 +155,7 @@ font-size: 21px;
     border-radius: 22px;
     background-color: #ff9d9e;
     box-shadow: inset 2px 2px 3px rgba(255, 255, 255, .6), inset -2px -2px 3px rgba(0, 0, 0, .6);
-    position: relative;
-    bottom: -32px;
-}
 
-
-.btn-valider:hover {
-    background-color: rgba(255, 0, 0, 1);
 }
 
 .btn-valider:active {
